minor refactoring

+++ src/tools/knifetool.cpp
@@ -118,10 +118,10 @@ void KnifeTool::draw(Painter& renderer) const
   SelectPointsBaseTool::draw(renderer);
   if (m_is_cutting) {
     renderer.painter->setPen(ui_color(HandleStatus::Active, "Handle", "foreground"));
-    renderer.painter->drawLine(m_mouse_press_pos.x,
-                               m_mouse_press_pos.y,
-                               m_mouse_move_pos.x,
-                               m_mouse_move_pos.y);
+    renderer.painter->drawLine(static_cast<int>(m_mouse_press_pos.x),
+                               static_cast<int>(m_mouse_press_pos.y),
+                               static_cast<int>(m_mouse_move_pos.x),
+                               static_cast<int>(m_mouse_move_pos.y));
     for (const Point& p : m_points) {
       QPen pen;
       pen.setColor(Qt::white);
